2.3 RecyclerView - added in branch master with 2 errors

+++ app/src/main/java/ru/netology/nmedia/NumberFormat.kt
@@ -1,12 +1,12 @@
 package ru.netology.nmedia
 
-    fun numberFormat(number: Int): String {
+    fun numberFormat(number: Long): String {
         return when {
             number < 1_000 -> number.toString()
             number in 1_000..9_999 -> {
                 val beforeDecimalPoint = number / 1_000
                 val afterDecimalPart = (number % 1_000) / 100
-                if (afterDecimalPart == 0) {
+                if (afterDecimalPart == 0.toLong()) {
                     String.format("%dK", beforeDecimalPoint)
                 } else {
                     String.format("%d,%dK", beforeDecimalPoint, afterDecimalPart)
@@ -21,7 +21,7 @@ package ru.netology.nmedia
             number in 1_000_000..99_999_999 -> {
                 val beforeDecimalPoint = number / 1_000_000
                 val afterDecimalPart = (number % 1_000_000) / 100_000
-                if (afterDecimalPart == 0) {
+                if (afterDecimalPart == 0.toLong()) {
                     String.format("%dМ", beforeDecimalPoint)
                 } else {
                     String.format("%d,%dМ", beforeDecimalPoint, afterDecimalPart)
